+++ capture/noworkflow/now/models/definition.py
@@ -1,0 +1,83 @@
+# Copyright (c) 2016 Universidade Federal Fluminense (UFF)
+# Copyright (c) 2016 Polytechnic Institute of New York University.
+# This file is part of noWorkflow.
+# Please, consult the license terms in the LICENSE file.
+"""Definition Object"""
+from __future__ import (absolute_import, print_function,
+                        division, unicode_literals)
+
+from collections import OrderedDict
+
+from future.utils import viewkeys
+
+from ..persistence.models.base import Model, proxy_gen
+from ..persistence.models.trial import Trial
+from .graphs.definition_graph import DefinitionGraph
+
+
+class Definition(Model):
+    """This model represents a definition of trials
+    Initialize it by passing a trials id:
+        definition = Definition(1)
+
+    There are four visualization modes for the graph:
+        tree: activation tree without any filters
+            definition.graph.mode = 0
+        no match: tree transformed into a graph by the addition of sequence and
+                  return edges and removal of intermediate call edges
+            definition.graph.mode = 1
+        exact match: calls are only combined when all the sub-call match
+            definition.graph.mode = 2
+        namesapce: calls are combined without considering the sub-calls
+            definition.graph.mode = 3
+
+
+    You can change the graph width and height by the variables:
+        definition.graph.width = 600
+        definition.graph.height = 400
+    """
+
+    __modelname__ = "Definition"
+
+    DEFAULT = {
+        "graph.width": 500,
+        "graph.height": 500,
+        "graph.mode": 0,
+        "graph.use_cache": False
+    }
+
+    REPLACE = {
+        "graph_width": "graph.width",
+        "graph_height": "graph.height",
+        "graph_mode": "graph.mode",
+        "graph_use_cache": "graph.use_cache"
+    }
+
+    def __init__(self, trial_ref, **kwargs):
+        super(Definition, self).__init__(trial_ref, **kwargs)
+        self.trial = Trial(trial_ref)
+        self.graph = DefinitionGraph(self)
+        self.initialize_default(kwargs)
+
+    @property
+    def modules(self):
+        """Diff modules from trials"""
+        return self.trial.modules
+
+    @property
+    def environment(self):
+        """Diff environment variables"""
+        return self.trial.environment_attrs
+
+    @property
+    def file_accesses(self):
+        """Diff file accesses"""
+        return self.trial.file_accesses
+
+    def _ipython_display_(self):
+        from IPython.display import display
+        bundle = {
+            'application/noworkflow.trial+json': self._modes[self.mode]()[1],
+            'text/plain': 'Trial {}'.format(self.trial.id),
+        }
+        display(bundle, raw=True)
