+++ capture/noworkflow/now/persistence/models/trial.py
@@ -1326,80 +1326,6 @@ class Trial(AlchemyProxy):
             return True
         return self.script == script
 
-    def code_component_definition(self):
-        """Return a code component definition"""
-        from .code_component import CodeComponent
-        return relational.session.query(CodeComponent.m).filter((
-            (CodeComponent.m.trial_id == self.id) &
-            (CodeComponent.m.type != "syntax")
-        )).all()
-
-    def composition_definition(self):
-        """Return a composition definition"""
-        from .composition import Composition
-        return relational.session.query(Composition.m).filter((
-            (Composition.m.trial_id == self.id) &
-            (Composition.m.type != '*op_pos')
-        )).all()
-
-    def label_def(self):
-        "Return the label needed for the trial's definition multi-name node."
-        from .code_component import CodeComponent
-        from .composition import Composition
-        def_id = relational.session.query(CodeComponent.m.id).join(
-            Composition.m,
-            (CodeComponent.m.id == Composition.m.part_id) &
-            (CodeComponent.m.trial_id == Composition.m.trial_id)
-        ).filter(
-            (CodeComponent.m.trial_id == self.id) &
-            (CodeComponent.m.type == 'function_def') | (CodeComponent.m.type == 'class_def')
-        ).subquery()
-
-        labels = relational.session.query(
-            CodeComponent.m.name,
-            CodeComponent.m.type,
-            Composition.m.whole_id,
-            Composition.m.position
-        ).join(
-            Composition.m,
-            (CodeComponent.m.id == Composition.m.part_id) &
-            (CodeComponent.m.trial_id == Composition.m.trial_id)
-        ).filter(
-            (CodeComponent.m.trial_id == self.id) &
-            (CodeComponent.m.type != 'syntax') &
-            Composition.m.whole_id.in_(def_id)
-        ).all()
-
-        return [{
-            'name': def_.name,
-            'type': def_.type,
-            'whole_id': def_.whole_id,
-            'position': def_.position}
-            for def_ in labels]
-
-    def definition(self):
-        """Return an AST representation of trial's definition"""
-        from ...models.node_definition import construct_node, construct_relationship
-        code_components = self.code_component_definition()
-        compositions = self.composition_definition()
-        node_dict = {}
-        for component in code_components:
-            node = construct_node(component, self.label_def())
-            if node is None:
-                continue
-            node_dict[component.id] = node
-
-        for composition in compositions:
-            whole_node = node_dict.get(composition.whole_id)
-            part_node = node_dict.get(composition.part_id)
-
-            if whole_node is None or part_node is None:
-                if composition.extra is None:
-                    continue
-            construct_relationship(whole_node, part_node, composition)
-
-        return node_dict[1]
-
     @property
     def str_start(self):
         """Return start date as string"""
